+++ src/main/java/org/synyx/urlaubsverwaltung/account/AccountViewController.java
@@ -63,7 +63,7 @@ public class AccountViewController implements HasLaunchpad {
 
     @PreAuthorize(IS_OFFICE)
     @GetMapping("/person/{personId}/account")
-    public String editAccount(@PathVariable("personId") Integer personId,
+    public String editAccount(@PathVariable("personId") Long personId,
                               @RequestParam(value = "year", required = false) Integer year, Model model)
         throws UnknownPersonException {
 
@@ -88,7 +88,7 @@ public class AccountViewController implements HasLaunchpad {
 
     @PreAuthorize(IS_OFFICE)
     @PostMapping("/person/{personId}/account")
-    public String updateAccount(@PathVariable("personId") Integer personId,
+    public String updateAccount(@PathVariable("personId") Long personId,
                                 @ModelAttribute("account") AccountForm accountForm, Errors errors, Model model,
                                 RedirectAttributes redirectAttributes) throws UnknownPersonException {
