+++ src/main/java/org/synyx/urlaubsverwaltung/application/statistics/ApplicationForLeaveStatisticsService.java
@@ -60,7 +60,7 @@ class ApplicationForLeaveStatisticsService {
         final Pageable pageable = pageableSearchQuery.getPageable();
         final List<VacationType> activeVacationTypes = vacationTypeService.getActiveVacationTypes();
         final Page<Person> relevantPersonsPage = getAllRelevantPersons(person, pageableSearchQuery);
-        final List<Integer> personIdValues = relevantPersonsPage.getContent().stream().map(Person::getId).collect(toList());
+        final List<Long> personIdValues = relevantPersonsPage.getContent().stream().map(Person::getId).collect(toList());
         final Map<PersonId, PersonBasedata> basedataByPersonId = personBasedataService.getBasedataByPersonId(personIdValues);
 
         final Collection<ApplicationForLeaveStatistics> statisticsCollection = applicationForLeaveStatisticsBuilder
