+++ src/main/java/org/synyx/urlaubsverwaltung/notification/UserNotificationSettingsRepository.java
@@ -2,5 +2,5 @@ package org.synyx.urlaubsverwaltung.notification;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-interface UserNotificationSettingsRepository extends JpaRepository<UserNotificationSettingsEntity, Integer> {
+interface UserNotificationSettingsRepository extends JpaRepository<UserNotificationSettingsEntity, Long> {
 }
