+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/web/OvertimeListRecordDto.java
@@ -14,7 +14,7 @@ public final class OvertimeListRecordDto {
         ABSENCE,
     }
 
-    private final Integer id;
+    private final Long id;
     private final LocalDate startDate;
     private final LocalDate endDate;
     private final Duration duration;
@@ -25,7 +25,7 @@ public final class OvertimeListRecordDto {
     private final String type;
     private final boolean isAllowedToEdit;
 
-    OvertimeListRecordDto(Integer id, LocalDate startDate, LocalDate endDate, Duration duration, Map<Integer, Duration> durationByYear, Duration sum, String status, String color, String type, boolean isAllowedToEdit) {
+    OvertimeListRecordDto(Long id, LocalDate startDate, LocalDate endDate, Duration duration, Map<Integer, Duration> durationByYear, Duration sum, String status, String color, String type, boolean isAllowedToEdit) {
         this.id = id;
         this.startDate = startDate;
         this.endDate = endDate;
@@ -43,7 +43,7 @@ public final class OvertimeListRecordDto {
             durationByYear, sum, overtimeListRecordDto.getStatus(), overtimeListRecordDto.getColor(), overtimeListRecordDto.getType(), overtimeListRecordDto.isAllowedToEdit);
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
