+++ src/main/java/org/synyx/urlaubsverwaltung/overtime/web/UnknownOvertimeException.java
@@ -6,7 +6,7 @@ import org.synyx.urlaubsverwaltung.web.AbstractNoResultFoundException;
  * Thrown in case no overtime record found for a certain ID.
  */
 public class UnknownOvertimeException extends AbstractNoResultFoundException {
-    UnknownOvertimeException(Integer id) {
+    UnknownOvertimeException(Long id) {
         super(id, "overtime");
     }
 }
